SGit views name refactor
